requirements.txt
fastapi==0.115.12
python-dotenv==1.1.0
requests==2.32.3
uvicorn==0.34.3
starlette==0.46.0